requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-openai
langchain-community
langgraph
chromadb
tiktoken
pypdf
python-dotenv
langchain-groq
sentence-transformers